+++ NEWS.md
@@ -1,3 +1,8 @@
+# ggdebug 0.1.1
+
+* Add `get_geom()` for determining the name of the Geom being used with a Stat
+    * Thanks to Brodie Gaslam for the assistance.
+
 # ggdebug 0.1.0
 
 * Initial release.